# docs/openapi/api-internal-pn-bff-pa-notifications.yaml
openapi: 3.0.3
info:
  title: PN BFF BE Microservice - Notifiche Mittenti
  description: Documentation APIs v1.0
  termsOfService: https://termofservice.it
  x-api-id: api-internal-pn-bff-pa
  x-summary: 'API di pn-bff esposte al Web ai mittenti per la visualizzazione e gestione delle notifiche'
  version: '1.0.0'
  contact:
    email: pn@pagopa.it
  license:
    name: Licenza di PN
    url: 'https://da-definire/'
servers:
  - url: https://webapi.notifichedigitali.it
    description: Ambiente di produzione
  - url: https://webapi.uat.notifichedigitali.it
    description: Ambiente di UAT
  - url: https://webapi.test.notifichedigitali.it
    description: Ambiente di test
  - url: https://webapi.dev.notifichedigitali.it
    description: Ambiente di sviluppo
tags:
  - name: NotificationSent
    description: >-
      Invocazioni utilizzabili dai mittenti per verificare lo stato delle richieste 
      di notifica inviate e delle notifiche accettate.

paths:
  '/bff/v1/notifications/sent':
    get:
      summary: 'Mittente: ricerca notifiche inviate'
      description: |-
        Questa operazione permette di ricercare le notifiche inviate.
      tags:
        - NotificationSent
      operationId: searchSentNotificationsV1
#      security:                                    # ONLY EXTERNAL
#       - bearerAuth: [ ]                           # ONLY EXTERNAL
      parameters:
        - $ref: './common-refs.yaml#/components/parameters/uidAuthFleet'                # NO EXTERNAL
        - $ref: './common-refs.yaml#/components/parameters/cxTypeAuthFleet'             # NO EXTERNAL
        - $ref: './common-refs.yaml#/components/parameters/cxIdAuthFleet'               # NO EXTERNAL
        - $ref: './common-refs.yaml#/components/parameters/cxGroupsAuthFleet'           # NO EXTERNAL
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffNotificationSearchStartDate'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffNotificationSearchEndDate'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffNotificationSearchRecipientId'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffNotificationSearchStatus'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffNotificationSearchSubject'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffSearchIun'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffNotificationSearchPageSize'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffNotificationSearchNextPagesKey'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './notification-schemas/notification.yaml#/components/schemas/BffNotificationsResponse'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: './common-refs.yaml#/components/schemas/Problem'
        '500':
          description: Internal error
          content:
            application/problem+json:
              schema:
                $ref: './common-refs.yaml#/components/schemas/Problem'

  '/bff/v1/notifications/sent/{iun}':
    get:
      summary: 'Mittente: lettura dettagli notifica'
      description: |-
        Questa operazione permette di leggere tutti i dettagli di una notifica accettata.
      tags:
        - NotificationSent
      operationId: getSentNotificationV1
#      security:                                      # ONLY EXTERNAL
#        - bearerAuth: [ ]                            # ONLY EXTERNAL
      parameters:
        - $ref: './common-refs.yaml#/components/parameters/uidAuthFleet'                # NO EXTERNAL
        - $ref: './common-refs.yaml#/components/parameters/cxTypeAuthFleet'             # NO EXTERNAL
        - $ref: './common-refs.yaml#/components/parameters/cxIdAuthFleet'               # NO EXTERNAL
        - $ref: './common-refs.yaml#/components/parameters/cxGroupsAuthFleet'           # NO EXTERNAL
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffPathIun'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './notification-schemas/notification.yaml#/components/schemas/BffFullNotificationV1'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: './common-refs.yaml#/components/schemas/Problem'
        '500':
          description: Internal error
          content:
            application/problem+json:
              schema:
                $ref: './common-refs.yaml#/components/schemas/Problem'

  '/bff/v1/notifications/sent/{iun}/documents':
    get:
      summary: |-
        Questa operazione permette di scaricare qualsiasi documento legato alla notifica.
      tags:
        - NotificationSent
      operationId: getSentNotificationDocumentV1
#      security:                                      # ONLY EXTERNAL
#        - bearerAuth: [ ]                            # ONLY EXTERNAL
      parameters:
        - $ref: './common-refs.yaml#/components/parameters/uidAuthFleet'                # NO EXTERNAL
        - $ref: './common-refs.yaml#/components/parameters/cxTypeAuthFleet'             # NO EXTERNAL
        - $ref: './common-refs.yaml#/components/parameters/cxIdAuthFleet'               # NO EXTERNAL
        - $ref: './common-refs.yaml#/components/parameters/cxGroupsAuthFleet'           # NO EXTERNAL
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffPathIun'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffQueryDocumentType'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffQueryDocumentId'
        - $ref: './notification-schemas/parameters-notification.yaml#/components/parameters/BffQueryLegalFactCategory'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "./notification-schemas/notification-documents.yaml#/components/schemas/BffDocumentDownloadMetadataResponse"
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: './common-refs.yaml#/components/schemas/Problem'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: './common-refs.yaml#/components/schemas/Problem'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: './common-refs.yaml#/components/schemas/Problem'